requirements.txt
fastapi
uvicorn
pillow
imageio
onnxruntime